Make system tests more convenient
  * manage run directory in the system_test dir
  * correct VLOG severities and improve log messages
  * compute objective function
  * a test fails on scheduler error
  * introduce default scheduler and run directory

+++ system_tests/internals/test.py
@@ -4,22 +4,30 @@ import traceback
 import logging
 import subprocess
 import signal
+import json
+from contextlib import contextmanager
 
 from internals import timer
+from internals import utils
 from internals.loop import EventLoop
+from internals.objective_function.history import History
+from internals.objective_function.objective_function import ObjectiveFunction
 from internals.state import State
 from internals.story import Story
 
 LOGGER = logging.getLogger('test_runner')
 
 LSS_ASSIGNMENTS_DIR = 'assignments'
+LSS_INPUT_DIR = 'input'
 LSS_INPUT_NAME = 'input'
 
+
 class Test:
-    def __init__(self, test_data_path, lss_input_dir, lss_executable_path, log_dir):
-        self.__has_failed = False
-        self.__lss_input_path = os.path.join(lss_input_dir, LSS_INPUT_NAME)
-        self.__lss_assignments_dir = os.path.join(lss_input_dir, LSS_ASSIGNMENTS_DIR)
+    def __init__(self, test_data_path, run_dir, lss_executable_path, log_dir):
+        self.has_failed = False
+        self.__lss_input_dir = os.path.join(run_dir, LSS_INPUT_DIR)
+        self.__lss_input_path = os.path.join(self.__lss_input_dir, LSS_INPUT_NAME)
+        self.__lss_assignments_dir = os.path.join(run_dir, LSS_ASSIGNMENTS_DIR)
         self.__lss_executable_path = lss_executable_path
         self.__log_dir = log_dir
         self.__state = None
@@ -31,25 +39,40 @@ class Test:
         run_lss_command = "{executable} --input={input} --assignments={assignments}".format(**{
             "executable": self.__lss_executable_path,
             "input": os.path.abspath(self.__lss_input_path),
-            "assignments": os.path.abspath(self.__lss_assignments_dir)
-        })
+            "assignments": os.path.abspath(self.__lss_assignments_dir)})
         scheduler = subprocess.Popen(run_lss_command, shell=True)
         try:
             EventLoop(self.__story, self.__state).run()
-            self.__state.write_history(self.__log_dir)
+            self.__process_result()
         except Exception as e:
             traceback.print_exc()
             LOGGER.warning(e)
-            self.__has_failed = True
+            self.has_failed = True
         finally:
-            scheduler.send_signal(signal.SIGINT)
-
-    def has_failed(self):
-        return self.__has_failed
+            if scheduler.poll():  # Scheduler has stopped
+                self.has_failed = True
+            else:
+                scheduler.send_signal(signal.SIGINT)
 
     def __prepare_for_running(self):
+        utils.make_empty_dir(self.__lss_assignments_dir)
+        utils.make_empty_dir(self.__lss_input_dir)
         self.__state = State(self.__story,
                              self.__lss_input_path,
                              self.__lss_assignments_dir)
-        self.__has_failed = False
+        self.has_failed = False
         timer.setup(self.__story.get_raw('mint'))
+
+    def __process_result(self):
+        raw_history = self.__state.gather_history()
+        self.__write_history_to_files(raw_history)
+
+        history = History(raw_history, self.__story.get_raw('jobs'))
+        self.result = ObjectiveFunction(history).compute()
+
+    def __write_history_to_files(self, raw_history):
+        log_file = os.path.join(self.__log_dir, 'history')
+        with open(log_file, 'wb') as f:
+            pickle.dump(raw_history, f)
+        with open(log_file + '.json', 'wt') as f:
+            json.dump(raw_history, f, indent=2)
